Configuration: necessary compiler flag added for file in CentOS 7
Until now, Maneage failed to build the 'file' program on at least one
machine running CentOS version 7.9.2009 (with Linux kernel 3.10.0, and GCC
4.8.5) when running './project configure'.

With this commit (as suggested by the error message issued by the
compiler), the '-std=c99' is passed to the compiler in the 'file' recipe
(within 'basic.mk'). This flag puts the compiler in C99 mode, which forces
it to compile code according to the 1999 edition of the C standard. This
was necessary for older versions of GCC (for example GCC 4.8.5 was released
in June 2015); hence why others hadn't reported this issue until now.

After this fix, File compiles succesfully on such systems; without causing
any problem with newer GCC versions (tested in GCC 12.1.0).

This issue was solved with the help of Pedram Ashofte Ardakani and Mohammad
Akhlaghi.