Initialize: including *.conf files from the preparation phase
Until now, when some configuration files were generated in the preparation
phase, only the auto-generated (in the preparation phase) Makefiles (with
the suffix '*.mk') were always included in 'initialize.mk'. As a
consequence, if there were any configuration files (with suffix '*.conf'),
they would not be automatically added, and it was necessary to manually
include them.

Since auto-generated configuration files are also one common output of the
preparation phase of a project, it is better to include them automatically.

With this commit, the '*.conf' configuration files generated in the
preparation phase are added by 'initialize.mk' automatically (if
necessary).

In the process, the comments in the final target of
'reproduce/analysis/make/prepare.mk' were updated to be more clear.