Configuration: astropy added as a dependency of astrometry.net
Until now, astrometry.net didn't explicitly depend on Astropy! However,
astrometry.net does depend on it (or pyfits) and will crash later when
running if astropy or pyfits aren't installed. We hadn't noticed this until
now because of the data reduction projects we had built Astropy
independently of astrometry.net! We noticed this bug in a project that
didn't use astropy!

With this commit, astropy is built as a dependency of astrometry.net and
afterwards (during the analysis), astrometry.net was able to run without
any crash.

This bug was found with the help of Zohreh Ghaffari.