Maneage'd A&A (Astronomy and Astrophysics) journal style
Until this commit, the LaTeX style and settings of the core Maneage branch
were very generic and customizing for each separate journal required some
time to prepare the LaTeX style.

With this commit, a first attempt at customization of Maneage for the LaTeX
styles of the A&A journal. Note that EDP Science (the publisher of A&A) has
only claimed copyright in their style files, but hasn't actually specified
a license. Given that they have no objection to arXiv or Overleaf freely
distributing their style files, there is no problem is releasing these
files within this branch of Maneage also (only for people who want to use
their style to publish in their journal).