Distribution tarball: edited for A&A's BibTeX format
Until now, the rule to build the contents of 'make dist' assumed the
default BibLaTeX peculiarities of the main Maneage branch. However, A&A
doesn't use that format and it was problematic for people submitting to
that journal.

With this commit, the necessary corrections have been made to the rule.