make dist: allows updating of references.tex in tarball
Until this commit, when we wanted to build the project's PDF within the
tarball (outside of Maneage) and it was necessary to update
'references.tex', the update would not be applied in the PDF. This was
because BibTeX only allows a single '.bib' file that was hard-coded in the
distribution tarball.

With this commit, the temporary Makefile that is also built in the
distribution tarball also creates 'references.bib' when necessary. Thus
fixing the issue above.